# server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Bible API
  description: |
    Complete Bible API with Role-Based Access Control (RBAC), multi-version support, RSS feeds, user management, file locations, and backup system.
    
    ## Role-Based Access Control
    - **Admin**: Full system access with all permissions
    - **User**: Standard user with Bible reading and format access
    - **Viewer**: Read-only access to Bible content
    - **Editor**: Can edit content but not manage users
    
    ## Authentication
    Use Bearer token authentication. Login to get your JWT token.
    
    ## Rate Limiting
    - General API: 100 requests per 15 minutes
    - Auth endpoints: 5 attempts per 15 minutes
    
    ## File Locations
    Users can configure local drive paths where TXT and XML files will be saved automatically when accessing verses.
  version: 3.0.0
  contact:
    name: API Support
    email: support@bibleapi.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.bibleapi.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /user/me:
    get:
      summary: Get complete user information
      description: Returns user profile, role, permissions, assigned Bible versions, and available formats
      tags: [User Information]
      security:
        - bearerAuth: []
      responses:
        200:
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      user_info:
                        type: object
                      bible_versions:
                        type: array
                      available_formats:
                        type: array
                      summary:
                        type: object
        401:
          description: Authentication required
...
  /user/me/formats:
    get:
      summary: Get user's available formats with access links
      description: Returns all formats user has access to with direct URLs
      tags: [User Information]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Formats retrieved successfully
        401:
          description: Authentication required
  /user/me/versions:
    get:
      summary: Get user's assigned Bible versions
      description: Returns all Bible versions user has access to with API endpoints
      tags: [User Information]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Bible versions retrieved successfully
        401:
          description: Authentication required
  /user/me/dashboard:
    get:
      summary: Get user dashboard summary
      description: Returns user statistics, recent verses, and quick access links
      tags: [User Information]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Dashboard data retrieved successfully
        401:
          description: Authentication required

  # ==================== FILE LOCATIONS ====================
...
  /bibleshow/json:
    get:
      summary: Get JSON format
      description: Returns JSON format of the latest verses
      tags: [Formats]
      security:
        - bearerAuth: []
      responses:
        200:
          description: JSON formatted data
          content:
            application/json:
              schema:
                type: object
        401:
          description: Authentication required
        403:
          description: Format access denied

  # ==================== ROLES & PERMISSIONS (ADMIN ONLY) ====================
...
  /admin/users:
    get:
      summary: Get all users
      description: Get list of all users with their roles (Admin only)
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
        403:
          description: Admin access required
  /admin/users/{userId}:
    get:
      summary: Get user by ID
      description: Get detailed information about a specific user (Admin only)
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: User details
        403:
          description: Admin access required
        404:
          description: User not found

    patch:
      summary: Update user
      description: Update user information (Admin only)
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                full_name:
                  type: string
                is_active:
                  type: boolean
      responses:
        200:
          description: User updated
        403:
          description: Admin access required
        404:
          description: User not found
...
  /user/create:
    post:
      summary: Create new user
      description: Create a new user account (Admin only)
      tags: [User Management]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - password
              properties:
                user_id:
                  type: string
                  minLength: 3
                  maxLength: 50
                password:
                  type: string
                  minLength: 4
                email:
                  type: string
                  format: email
                full_name:
                  type: string
                role_code:
                  type: string
                  enum: [admin, user, viewer, editor]
                  default: user
      responses:
        200:
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
        400:
          description: Validation error or user already exists
        403:
          description: Admin access required
  /user/assign-formats:
    post:
      summary: Assign formats to user
      description: Assign format permissions to a user (Admin only)
      tags: [User Management]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - format_codes
              properties:
                user_id:
                  type: string
                format_codes:
                  type: array
                  items:
                    type: string
                    enum: [RSS_VERSE, RSS_REFERENCE, XML_BIBLESHOW, JSON_BIBLE, TXT_VERSE]
      responses:
        200:
          description: Formats assigned
        400:
          description: Invalid request
        403:
          description: Admin access required
        404:
          description: User not found
  /user/assign-versions:
    post:
      summary: Assign Bible versions to user
      description: Assign Bible version access to a user (Admin only)
      tags: [User Management]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - version_codes
              properties:
                user_id:
                  type: string
                version_codes:
                  type: array
                  items:
                    type: string
                    enum: [KJV, TAM]
                default_version:
                  type: string
      responses:
        200:
          description: Versions assigned
        400:
          description: Invalid request
        403:
          description: Admin access required
        404:
          description: User not found
  /user/create-full:
    post:
      summary: Create user with formats and versions
      description: Create a new user with formats and versions assigned (Admin only)
      tags: [User Management]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - password
              properties:
                user_id:
                  type: string
                password:
                  type: string
                email:
                  type: string
                full_name:
                  type: string
                role_code:
                  type: string
                  enum: [admin, user, viewer, editor]
                format_codes:
                  type: array
                  items:
                    type: string
                version_codes:
                  type: array
                  items:
                    type: string
                default_version:
                  type: string
      responses:
        200:
          description: User created with assignments
        400:
          description: Invalid request
        403:
          description: Admin access required
...